Move all I/O operations into a state monad.
Because reads in certain situations (i.e. the PPU) can change state, the
previous method of (mem-read) returning a byte and (mem-write) returning
a new CPU state doesn't really work.

Instead both are wrapped in a state monad (with some nice helper
functions) that returns a new pair [val-read-or-written new-state].