Refactor away from mounting memory devices.
Instead use a memory bus object to coordinate which I/O operations apply
to which devices. This will require having a separate memory bus for the
PPU and the CPU, but it vastly simplifies and speeds up the I/O API.

The mounting idea was clever, but just didn't work as well as the simple
solution. Shame on me.